+++ finch/__init__.py
@@ -3,5 +3,3 @@
 """Top-level package for Finch."""
 
 from .__version__ import __author__, __email__, __version__  # noqa: F401
-
-from .wsgi import application  # noqa: F401